+++ src/main/java/io/github/eappezo/soundary/services/music/application/authentication/MusicPlatformAuthenticationManagerImpl.java
@@ -1,6 +1,6 @@
 package io.github.eappezo.soundary.services.music.application.authentication;
 
-import io.github.eappezo.soundary.services.music.application.PlatformAccessToken;
+import io.github.eappezo.soundary.services.music.domain.PlatformAccessToken;
 import io.github.eappezo.soundary.services.music.domain.MusicPlatform;
 import io.github.eappezo.soundary.services.music.domain.MusicPlatformAuthenticationManager;
 import lombok.RequiredArgsConstructor;
