+++ src/main/java/io/github/eappezo/soundary/services/music/application/authentication/MusicPlatformOAuthGateway.java
@@ -1,6 +1,6 @@
 package io.github.eappezo.soundary.services.music.application.authentication;
 
-import io.github.eappezo.soundary.services.music.application.PlatformAccessToken;
+import io.github.eappezo.soundary.services.music.domain.PlatformAccessToken;
 import io.github.eappezo.soundary.services.music.domain.MusicPlatform;
 
 public interface MusicPlatformOAuthGateway {
