+++ src/main/java/io/github/eappezo/soundary/services/music/domain/MusicPlatformAuthenticationManager.java
@@ -1,7 +1,5 @@
 package io.github.eappezo.soundary.services.music.domain;
 
-import io.github.eappezo.soundary.services.music.application.PlatformAccessToken;
-
 public interface MusicPlatformAuthenticationManager {
 
     PlatformAccessToken getAccessToken(MusicPlatform platform);
