+++ src/main/java/io/github/eappezo/soundary/services/music/domain/PlatformAccessToken.java
@@ -1,6 +1,4 @@
-package io.github.eappezo.soundary.services.music.application;
-
-import io.github.eappezo.soundary.services.music.domain.MusicPlatform;
+package io.github.eappezo.soundary.services.music.domain;
 
 public record PlatformAccessToken(
         MusicPlatform platform,
