+++ src/main/java/io/github/eappezo/soundary/services/music/infrastructure/gateway/SpotifyOAuthGateway.java
@@ -3,7 +3,7 @@ package io.github.eappezo.soundary.services.music.infrastructure.gateway;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import io.github.eappezo.soundary.services.authentication.domain.exception.AuthenticationFailedException;
-import io.github.eappezo.soundary.services.music.application.PlatformAccessToken;
+import io.github.eappezo.soundary.services.music.domain.PlatformAccessToken;
 import io.github.eappezo.soundary.services.music.application.authentication.MusicPlatformOAuthGateway;
 import io.github.eappezo.soundary.services.music.domain.MusicPlatform;
 import io.github.eappezo.soundary.services.music.infrastructure.MusicPlatformOAuthConfig;
