+++ src/main/java/io/github/eappezo/soundary/services/music/infrastructure/persistence/dao/support/QSharedMusicRetrieveSupport.java
@@ -1,9 +1,11 @@
 package io.github.eappezo.soundary.services.music.infrastructure.persistence.dao.support;
 
+import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class SharedMusicRetrieveSupportImpl {
+public class QSharedMusicRetrieveSupport {
+    private final JPAQueryFactory jpaQueryFactory;
 }
